# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Real-time multiplayer fitness backend (Aviron-inspired). See context/project-overview.md for the full design.",
        "title": "Aviron Backend API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Exchanges email and password for a JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Log in",
                "parameters": [
                    {
                        "description": "Login payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_auth.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_auth.loginResponse"
                        }
                    },
                    "401": {
                        "description": "error: invalid_credentials",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Creates a new user account with an email, password, and display name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Registration payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_auth.registerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_auth.registerResponse"
                        }
                    },
                    "400": {
                        "description": "field-keyed validation errors",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "error: email_taken",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/healthz": {
            "get": {
                "description": "Pings the database connection pool and reports service health",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "status: ok",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "status: db_unreachable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/races": {
            "post": {
                "description": "Creates a new typing race with a name and target word count",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "races"
                ],
                "summary": "Create a race",
                "parameters": [
                    {
                        "description": "Create race payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_race.createRaceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_race.createRaceResponse"
                        }
                    },
                    "400": {
                        "description": "field-keyed validation errors",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "error: unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/races/{id}": {
            "get": {
                "description": "Returns a race's current status and participant list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "races"
                ],
                "summary": "Get race status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Race ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_race.raceStatusResponse"
                        }
                    },
                    "400": {
                        "description": "error: invalid_race_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "error: unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "error: race_not_found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/races/{id}/join": {
            "post": {
                "description": "Joins an existing race as a participant, returning a per-race session token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "races"
                ],
                "summary": "Join a race",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Race ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_race.joinRaceResponse"
                        }
                    },
                    "400": {
                        "description": "error: invalid_race_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "error: unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "error: race_not_found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "error: already_joined | race_not_pending | race_full",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/races/{id}/leave": {
            "post": {
                "description": "Removes the caller as a participant of a still-pending race. Once the race is active, quit via the WebSocket leave_race message instead.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "races"
                ],
                "summary": "Leave a race",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Race ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_race.leaveRaceResponse"
                        }
                    },
                    "400": {
                        "description": "error: invalid_race_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "error: unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "error: race_not_found | not_participant",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "error: race_not_pending",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/races/{id}/start": {
            "post": {
                "description": "Creator starts the race: generates the shared prompt text and flips status to active",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "races"
                ],
                "summary": "Start a race",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Race ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_race.startRaceResponse"
                        }
                    },
                    "400": {
                        "description": "error: invalid_race_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "error: unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "error: forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "error: race_not_found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "error: race_not_pending",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/races/{id}/text": {
            "get": {
                "description": "Fetches the race's already-generated prompt text",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "races"
                ],
                "summary": "Get race prompt text",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Race ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_race.getRaceTextResponse"
                        }
                    },
                    "400": {
                        "description": "error: invalid_race_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "error: unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "error: race_not_found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "error: prompt_not_ready",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_auth.loginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "internal_auth.loginResponse": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "internal_auth.registerRequest": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "internal_auth.registerResponse": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "internal_race.createRaceRequest": {
            "type": "object",
            "properties": {
                "distance_meters": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "internal_race.createRaceResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "distance_meters": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "internal_race.getRaceTextResponse": {
            "type": "object",
            "properties": {
                "prompt_text": {
                    "type": "string"
                }
            }
        },
        "internal_race.joinRaceResponse": {
            "type": "object",
            "properties": {
                "race_id": {
                    "type": "string"
                },
                "session_token": {
                    "type": "string"
                }
            }
        },
        "internal_race.leaveRaceResponse": {
            "type": "object",
            "properties": {
                "race_id": {
                    "type": "string"
                }
            }
        },
        "internal_race.participantResponse": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "joined_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "internal_race.raceStatusResponse": {
            "type": "object",
            "properties": {
                "created_by": {
                    "type": "string"
                },
                "distance_meters": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "participants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_race.participantResponse"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "internal_race.startRaceResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "prompt_text": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}