# backend/internal/race/handler.go
package race

import (
	"context"
	"encoding/json"
	"errors"
	"net/http"
	"regexp"
	"time"

	"github.com/akkien/aviron/internal/httpx"
	"github.com/akkien/aviron/internal/middleware"
	"github.com/akkien/aviron/internal/room"
)

var uuidPattern = regexp.MustCompile(`^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$`)

func isValidUUID(s string) bool {
	return uuidPattern.MatchString(s)
}

type RaceHandler struct {
	svc      *RaceService
	registry *room.Registry
	// ctx is the process's root context, used (not r.Context()) when spawning
	// a room actor so it keeps running after the triggering HTTP request
	// returns and its own context is cancelled.
	ctx context.Context
}

func NewRaceHandler(svc *RaceService, registry *room.Registry, ctx context.Context) *RaceHandler {
	return &RaceHandler{svc: svc, registry: registry, ctx: ctx}
}

// Create godoc
// @Summary Create a race
// @Description Creates a new typing race with a name and target word count
// @Tags races
// @Accept json
// @Produce json
// @Param request body createRaceRequest true "Create race payload"
// @Success 201 {object} createRaceResponse
// @Failure 400 {object} map[string]interface{} "field-keyed validation errors"
// @Failure 401 {object} map[string]string "error: unauthorized"
// @Router /races [post]
func (h *RaceHandler) Create(w http.ResponseWriter, r *http.Request) {
	userID, ok := middleware.UserIDFromContext(r.Context())
	if !ok {
		httpx.WriteError(w, http.StatusUnauthorized, "unauthorized")
		return
	}

	var req createRaceRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		httpx.WriteError(w, http.StatusBadRequest, "invalid_body")
		return
	}

	created, fieldErrs, err := h.svc.CreateRace(r.Context(), req.Name, req.DistanceMeters, userID)
	if len(fieldErrs) > 0 {
		httpx.WriteJSON(w, http.StatusBadRequest, map[string]any{"errors": fieldErrs})
		return
	}
	if err != nil {
		httpx.WriteError(w, http.StatusInternalServerError, "internal_error")
		return
	}

	httpx.WriteJSON(w, http.StatusCreated, createRaceResponse{
		ID:             created.ID,
		Name:           created.Name,
		DistanceMeters: created.DistanceMeters,
		Status:         created.Status,
		CreatedBy:      created.CreatedBy,
		CreatedAt:      created.CreatedAt.Format(time.RFC3339),
	})
}

// Join godoc
// @Summary Join a race
// @Description Joins an existing race as a participant, returning a per-race session token
// @Tags races
// @Produce json
// @Param id path string true "Race ID"
// @Success 200 {object} joinRaceResponse
// @Failure 400 {object} map[string]string "error: invalid_race_id"
// @Failure 401 {object} map[string]string "error: unauthorized"
// @Failure 404 {object} map[string]string "error: race_not_found"
// @Failure 409 {object} map[string]string "error: already_joined | race_not_pending | race_full"
// @Router /races/{id}/join [post]
func (h *RaceHandler) Join(w http.ResponseWriter, r *http.Request) {
	userID, ok := middleware.UserIDFromContext(r.Context())
	if !ok {
		httpx.WriteError(w, http.StatusUnauthorized, "unauthorized")
		return
	}

	raceID := r.PathValue("id")
	if !isValidUUID(raceID) {
		httpx.WriteError(w, http.StatusBadRequest, "invalid_race_id")
		return
	}

	sessionToken, err := h.svc.JoinRace(r.Context(), raceID, userID)
	switch {
	case errors.Is(err, ErrRaceNotFound):
		httpx.WriteError(w, http.StatusNotFound, "race_not_found")
		return
	case errors.Is(err, ErrRaceNotPending):
		httpx.WriteError(w, http.StatusConflict, "race_not_pending")
		return
	case errors.Is(err, ErrAlreadyJoined):
		httpx.WriteError(w, http.StatusConflict, "already_joined")
		return
	case errors.Is(err, ErrRaceFull):
		httpx.WriteError(w, http.StatusConflict, "race_full")
		return
	case err != nil:
		httpx.WriteError(w, http.StatusInternalServerError, "internal_error")
		return
	}

	httpx.WriteJSON(w, http.StatusOK, joinRaceResponse{
		RaceID:       raceID,
		SessionToken: sessionToken,
	})
}

// Leave godoc
// @Summary Leave a race
// @Description Removes the caller as a participant of a still-pending race. Once the race is active, quit via the WebSocket leave_race message instead.
// @Tags races
// @Produce json
// @Param id path string true "Race ID"
// @Success 200 {object} leaveRaceResponse
// @Failure 400 {object} map[string]string "error: invalid_race_id"
// @Failure 401 {object} map[string]string "error: unauthorized"
// @Failure 404 {object} map[string]string "error: race_not_found | not_participant"
// @Failure 409 {object} map[string]string "error: race_not_pending"
// @Router /races/{id}/leave [post]
func (h *RaceHandler) Leave(w http.ResponseWriter, r *http.Request) {
	userID, ok := middleware.UserIDFromContext(r.Context())
	if !ok {
		httpx.WriteError(w, http.StatusUnauthorized, "unauthorized")
		return
	}

	raceID := r.PathValue("id")
	if !isValidUUID(raceID) {
		httpx.WriteError(w, http.StatusBadRequest, "invalid_race_id")
		return
	}

	err := h.svc.LeaveRace(r.Context(), raceID, userID)
	switch {
	case errors.Is(err, ErrRaceNotFound):
		httpx.WriteError(w, http.StatusNotFound, "race_not_found")
		return
	case errors.Is(err, ErrNotParticipant):
		httpx.WriteError(w, http.StatusNotFound, "not_participant")
		return
	case errors.Is(err, ErrRaceNotPending):
		httpx.WriteError(w, http.StatusConflict, "race_not_pending")
		return
	case err != nil:
		httpx.WriteError(w, http.StatusInternalServerError, "internal_error")
		return
	}

	httpx.WriteJSON(w, http.StatusOK, leaveRaceResponse{RaceID: raceID})
}

// Start godoc
// @Summary Start a race
// @Description Creator starts the race: generates the shared prompt text and flips status to active
// @Tags races
// @Produce json
// @Param id path string true "Race ID"
// @Success 200 {object} startRaceResponse
// @Failure 400 {object} map[string]string "error: invalid_race_id"
// @Failure 401 {object} map[string]string "error: unauthorized"
// @Failure 403 {object} map[string]string "error: forbidden"
// @Failure 404 {object} map[string]string "error: race_not_found"
// @Failure 409 {object} map[string]string "error: race_not_pending"
// @Router /races/{id}/start [post]
func (h *RaceHandler) Start(w http.ResponseWriter, r *http.Request) {
	userID, ok := middleware.UserIDFromContext(r.Context())
	if !ok {
		httpx.WriteError(w, http.StatusUnauthorized, "unauthorized")
		return
	}

	raceID := r.PathValue("id")
	if !isValidUUID(raceID) {
		httpx.WriteError(w, http.StatusBadRequest, "invalid_race_id")
		return
	}

	started, err := h.svc.StartRace(r.Context(), raceID, userID)
	switch {
	case errors.Is(err, ErrRaceNotFound):
		httpx.WriteError(w, http.StatusNotFound, "race_not_found")
		return
	case errors.Is(err, ErrNotCreator):
		httpx.WriteError(w, http.StatusForbidden, "forbidden")
		return
	case errors.Is(err, ErrRaceNotPending):
		httpx.WriteError(w, http.StatusConflict, "race_not_pending")
		return
	case err != nil:
		httpx.WriteError(w, http.StatusInternalServerError, "internal_error")
		return
	}

	promptText := ""
	if started.PromptText != nil {
		promptText = *started.PromptText
	}
	startedAt := ""
	if started.StartedAt != nil {
		startedAt = started.StartedAt.Format(time.RFC3339)
	}

	h.registry.Spawn(h.ctx, raceID, promptText, started.DistanceMeters, h.svc)

	httpx.WriteJSON(w, http.StatusOK, startRaceResponse{
		ID:         started.ID,
		Status:     started.Status,
		StartedAt:  startedAt,
		PromptText: promptText,
	})
}

// Text godoc
// @Summary Get race prompt text
// @Description Fetches the race's already-generated prompt text
// @Tags races
// @Produce json
// @Param id path string true "Race ID"
// @Success 200 {object} getRaceTextResponse
// @Failure 400 {object} map[string]string "error: invalid_race_id"
// @Failure 401 {object} map[string]string "error: unauthorized"
// @Failure 404 {object} map[string]string "error: race_not_found"
// @Failure 409 {object} map[string]string "error: prompt_not_ready"
// @Router /races/{id}/text [get]
func (h *RaceHandler) Text(w http.ResponseWriter, r *http.Request) {
	_, ok := middleware.UserIDFromContext(r.Context())
	if !ok {
		httpx.WriteError(w, http.StatusUnauthorized, "unauthorized")
		return
	}

	raceID := r.PathValue("id")
	if !isValidUUID(raceID) {
		httpx.WriteError(w, http.StatusBadRequest, "invalid_race_id")
		return
	}

	promptText, err := h.svc.GetRaceText(r.Context(), raceID)
	switch {
	case errors.Is(err, ErrRaceNotFound):
		httpx.WriteError(w, http.StatusNotFound, "race_not_found")
		return
	case errors.Is(err, ErrPromptNotReady):
		httpx.WriteError(w, http.StatusConflict, "prompt_not_ready")
		return
	case err != nil:
		httpx.WriteError(w, http.StatusInternalServerError, "internal_error")
		return
	}

	httpx.WriteJSON(w, http.StatusOK, getRaceTextResponse{PromptText: promptText})
}

// Status godoc
// @Summary Get race status
// @Description Returns a race's current status and participant list
// @Tags races
// @Produce json
// @Param id path string true "Race ID"
// @Success 200 {object} raceStatusResponse
// @Failure 400 {object} map[string]string "error: invalid_race_id"
// @Failure 401 {object} map[string]string "error: unauthorized"
// @Failure 404 {object} map[string]string "error: race_not_found"
// @Router /races/{id} [get]
func (h *RaceHandler) Status(w http.ResponseWriter, r *http.Request) {
	_, ok := middleware.UserIDFromContext(r.Context())
	if !ok {
		httpx.WriteError(w, http.StatusUnauthorized, "unauthorized")
		return
	}

	raceID := r.PathValue("id")
	if !isValidUUID(raceID) {
		httpx.WriteError(w, http.StatusBadRequest, "invalid_race_id")
		return
	}

	detail, err := h.svc.GetRaceDetail(r.Context(), raceID)
	switch {
	case errors.Is(err, ErrRaceNotFound):
		httpx.WriteError(w, http.StatusNotFound, "race_not_found")
		return
	case err != nil:
		httpx.WriteError(w, http.StatusInternalServerError, "internal_error")
		return
	}

	participants := make([]participantResponse, len(detail.Participants))
	for i, p := range detail.Participants {
		participants[i] = participantResponse{
			UserID:      p.UserID,
			DisplayName: p.DisplayName,
			JoinedAt:    p.JoinedAt.Format(time.RFC3339),
		}
	}

	httpx.WriteJSON(w, http.StatusOK, raceStatusResponse{
		ID:             detail.ID,
		Name:           detail.Name,
		DistanceMeters: detail.DistanceMeters,
		Status:         detail.Status,
		CreatedBy:      detail.CreatedBy,
		Participants:   participants,
	})
}